Otimizacao de imports, correcoes nos testes de produtos para lidar com valor double, teste de pedido e remocao de classes nao utilizadas

+++ app/src/test/java/br/com/appfastfood/produto/aplicacao/adaptadores/resposta/ProdutoRespostaTest.java
@@ -5,14 +5,12 @@ import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 
-import java.math.BigDecimal;
-
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class ProdutoRespostaTest {
 
     @Mock
-    private BigDecimal precoMock;
+    private Double precoMock;
 
     @InjectMocks
     private ProdutoRequisicao produtoRequisicao;
