+++ app/src/test/java/br/com/appfastfood/produto/servicos/ProdutoServicoImplTest.java
@@ -1,6 +1,6 @@
 package br.com.appfastfood.produto.servicos;
 
-import br.com.appfastfood.produto.dominio.modelos.*;
+import br.com.appfastfood.produto.dominio.modelos.Produto;
 import br.com.appfastfood.produto.dominio.repositorios.ProdutoRepositorio;
 import br.com.appfastfood.produto.dominio.servicos.adaptadores.ProdutoServicoImpl;
 import br.com.appfastfood.produto.dominio.vo.*;
@@ -8,7 +8,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 
-import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
@@ -31,7 +30,7 @@ class ProdutoServicoImplTest {
         // Arrange
         Produto produto = new Produto(1L,
                 new Nome("Exemplo de Produto"),
-                new Preco(BigDecimal.valueOf(10.99)),
+                new Preco(10.99D),
                 new UriImagem("https://exemplo.com/imagem.jpg"),
                 new Categoria("lanche").getCategoria(),
                 new Descricao("Descrição do exemplo de produto")
@@ -63,7 +62,7 @@ class ProdutoServicoImplTest {
         Produto produto = new Produto(
                 id,
                 new Nome("Exemplo de Produto"),
-                new Preco(BigDecimal.valueOf(10.99)),
+                new Preco(10.99D),
                 new UriImagem("https://exemplo.com/imagem.jpg"),
                 new Categoria("lanche").getCategoria(),
                 new Descricao("Descrição do exemplo de produto")
@@ -72,7 +71,7 @@ class ProdutoServicoImplTest {
         Produto produtoAlterado = new Produto(
             id,
                 new Nome("Exemplo de Produto Alterado"),
-                new Preco(BigDecimal.valueOf(10.99)),
+                new Preco(10.99D),
                 new UriImagem("https://exemplo.com/imagem.jpg"),
                 new Categoria("lanche").getCategoria(),
                 new Descricao("Descrição do exemplo de produto")
@@ -95,7 +94,7 @@ class ProdutoServicoImplTest {
         List<Produto> produtos = Arrays.asList(new Produto(
             1L,
                 new Nome("Exemplo de Produto Alterado"),
-                new Preco(BigDecimal.valueOf(10.99)),
+                new Preco(10.99D),
                 new UriImagem("https://exemplo.com/imagem.jpg"),
                 new Categoria("lanche").getCategoria(),
                 new Descricao("Descrição do exemplo de produto")
